Consolidate shared code for trigger attribute discovery
